fix glue/athena creation
