+++ infrastructure/aws-cdk/src/constructs/data-lake-construct.ts
@@ -42,180 +42,7 @@ const defaultProps: Partial<DataLakeConstructProps> = {};
 export class DataLakeConstruct extends Construct {
   public readonly gameEventsDatabase: glueCfn.CfnDatabase;
   public readonly rawEventsTable: glueCfn.CfnTable;
-
-  private createDefaultAthenaQueries(
-    databaseName: string,
-    tableName: string,
-    workgroupName: string
-  ) {
-    const queries = [
-      {
-        database: databaseName,
-        name: "LatestEventsQuery",
-        description: "Get latest events by event_timestamp",
-        workgroup: workgroupName,
-        query: `SELECT *, from_unixtime(event_timestamp, 'America/New_York') as event_timestamp_america_new_york
-                FROM "${databaseName}"."${tableName}"
-                ORDER BY event_timestamp_america_new_york DESC
-                LIMIT 10;`,
-      },
-      {
-        database: databaseName,
-        name: "TotalEventsQuery",
-        description: "Total events",
-        workgroup: workgroupName,
-        query: `SELECT application_id, count(DISTINCT event_id) as event_count 
-                FROM "${databaseName}"."${tableName}"
-                GROUP BY application_id`,
-      },
-      {
-        database: databaseName,
-        name: "TotalEventsMonthQuery",
-        description: "Total events over last month",
-        workgroup: workgroupName,
-        query: `WITH detail AS
-                (SELECT date_trunc('month', date(date_parse(CONCAT(year, '-', month, '-', day), '%Y-%m-%d'))) as event_month, * 
-                FROM "${databaseName}"."${tableName}") 
-                SELECT date_trunc('month', event_month) as month, application_id, count(DISTINCT event_id) as event_count 
-                FROM detail 
-                GROUP BY date_trunc('month', event_month), application_id`,
-      },
-      {
-        database: databaseName,
-        name: "TotalIapTransactionsLastMonth",
-        description: "Total IAP Transactions over the last month",
-        workgroup: workgroupName,
-        query: `WITH detail AS 
-                (SELECT date_trunc('month', date(date_parse(CONCAT(year, '-', month, '-', day),'%Y-%m-%d'))) as event_month,* 
-                FROM "${databaseName}"."${tableName}") 
-                SELECT date_trunc('month', event_month) as month, application_id, count(DISTINCT json_extract_scalar(event_data, '$.transaction_id')) as transaction_count 
-                FROM detail WHERE json_extract_scalar(event_data, '$.transaction_id') is NOT null 
-                AND event_type = 'iap_transaction'
-                GROUP BY date_trunc('month', event_month), application_id`,
-      },
-      {
-        database: databaseName,
-        name: "NewUsersLastMonth",
-        description: "New Users over the last month",
-        workgroup: workgroupName,
-        query: `WITH detail AS (
-                SELECT date_trunc('month', date(date_parse(CONCAT(year, '-', month, '-', day), '%Y-%m-%d'))) as event_month, *
-                FROM "${databaseName}"."${tableName}")
-                SELECT
-                date_trunc('month', event_month) as month,
-                count(*) as new_accounts
-                FROM detail
-                WHERE event_type = 'user_registration'
-                GROUP BY date_trunc('month', event_month);`,
-      },
-      {
-        database: databaseName,
-        name: "TotalPlaysByLevel",
-        description: "Total number of times each level has been played",
-        workgroup: workgroupName,
-        query: `SELECT
-                json_extract_scalar(event_data, '$.level_id') as level,
-                count(json_extract_scalar(event_data, '$.level_id')) as number_of_plays
-                FROM "${databaseName}"."${tableName}"
-                WHERE event_type = 'level_started'
-                GROUP BY json_extract_scalar(event_data, '$.level_id')
-                ORDER by json_extract_scalar(event_data, '$.level_id');`,
-      },
-      {
-        database: databaseName,
-        name: "TotalFailuresByLevel",
-        description: "Total number of failures on each level",
-        workgroup: workgroupName,
-        query: `SELECT
-                json_extract_scalar(event_data, '$.level_id') as level,
-                count(json_extract_scalar(event_data, '$.level_id')) as number_of_failures
-                FROM "${databaseName}"."${tableName}"
-                WHERE event_type='level_failed'
-                GROUP BY json_extract_scalar(event_data, '$.level_id')
-                ORDER by json_extract_scalar(event_data, '$.level_id');`,
-      },
-      {
-        database: databaseName,
-        name: "TotalCompletionsByLevel",
-        description: "Total number of completions on each level",
-        workgroup: workgroupName,
-        query: `SELECT
-                json_extract_scalar(event_data, '$.level_id') as level,
-                count(json_extract_scalar(event_data, '$.level_id')) as number_of_completions
-                FROM "${databaseName}"."${tableName}"
-                WHERE event_type='level_completed'
-                GROUP BY json_extract_scalar(event_data, '$.level_id')
-                ORDER by json_extract_scalar(event_data, '$.level_id');`,
-      },
-      {
-        database: databaseName,
-        name: "LevelCompletionRate",
-        description: "Rate of completion for each level",
-        workgroup: workgroupName,
-        query: `with t1 as
-                (SELECT json_extract_scalar(event_data, '$.level_id') as level, count(json_extract_scalar(event_data, '$.level_id')) as level_count 
-                FROM "${databaseName}"."${tableName}"
-                WHERE event_type='level_started' GROUP BY json_extract_scalar(event_data, '$.level_id') 
-                ),
-                t2 as
-                (SELECT json_extract_scalar(event_data, '$.level_id') as level, count(json_extract_scalar(event_data, '$.level_id')) as level_count 
-                FROM "${databaseName}"."${tableName}"
-                WHERE event_type='level_completed'GROUP BY json_extract_scalar(event_data, '$.level_id') 
-                )
-                select t2.level, (cast(t2.level_count AS DOUBLE) / (cast(t2.level_count AS DOUBLE) + cast(t1.level_count AS DOUBLE))) * 100 as level_completion_rate from 
-                t1 JOIN t2 ON t1.level = t2.level
-                ORDER by level;`,
-      },
-      {
-        database: databaseName,
-        name: "AverageUserSentimentPerDay",
-        description: "User sentiment score by day",
-        workgroup: workgroupName,
-        query: `SELECT
-                avg(CAST(json_extract_scalar(event_data, '$.user_rating') AS real)) AS average_user_rating, 
-                date(date_parse(CONCAT(year, '-', month, '-', day), '%Y-%m-%d')) as event_date
-                FROM "${databaseName}"."${tableName}"
-                WHERE json_extract_scalar(event_data, '$.user_rating') is not null
-                GROUP BY date(date_parse(CONCAT(year, '-', month, '-', day), '%Y-%m-%d'));`,
-      },
-      {
-        database: databaseName,
-        name: "UserReportedReasonsCount",
-        description: "Reasons users are being reported, grouped by reason code",
-        workgroup: workgroupName,
-        query: `SELECT count(json_extract_scalar(event_data, '$.report_reason')) as count_of_reports, json_extract_scalar(event_data, '$.report_reason') as report_reason
-                FROM "${databaseName}"."${tableName}"
-                GROUP BY json_extract_scalar(event_data, '$.report_reason')
-                ORDER BY json_extract_scalar(event_data, '$.report_reason') DESC;`,
-      },
-      {
-        database: databaseName,
-        name: "CTASCreateIcebergTables",
-        description: "Create table as (CTAS) from existing tables to iceberg",
-        workgroup: workgroupName,
-        query: `CREATE TABLE "${tableName}"."raw_events_iceberg"
-                WITH (table_type = 'ICEBERG',
-                    format = 'PARQUET', 
-                    location = 's3://your_bucket/', 
-                    is_external = false,
-                    partitioning = ARRAY['application_id', 'year', 'month', 'day'],
-                    vacuum_min_snapshots_to_keep = 10,
-                    vacuum_max_snapshot_age_seconds = 604800
-                ) 
-                AS SELECT * FROM "${databaseName}"."${tableName}";`,
-      },
-    ];
-
-    for (const query of queries) {
-      new athena.CfnNamedQuery(this, `NamedQuery-${query.name}`, {
-        database: query.database,
-        name: query.name,
-        workGroup: query.workgroup,
-        description: query.description,
-        queryString: query.query,
-      });
-    }
-  }
+  public readonly gameAnalyticsWorkgroup: athena.CfnWorkGroup;
 
   constructor(parent: Construct, name: string, props: DataLakeConstructProps) {
     super(parent, name);
@@ -229,10 +56,10 @@ export class DataLakeConstruct extends Construct {
       "GameEventsDatabase",
       {
         catalogId: cdk.Aws.ACCOUNT_ID,
+        databaseName: props.config.EVENTS_DATABASE,
         databaseInput: {
           description: `Database for game analytics events for stack: ${cdk.Aws.STACK_NAME}`,
-          locationUri: `s3://${props.analyticsBucket.bucketName}`,
-          name: props.config.EVENTS_DATABASE,
+          locationUri: props.analyticsBucket.bucketArn,
         },
       }
     );
@@ -243,7 +70,7 @@ export class DataLakeConstruct extends Construct {
       this,
       "GameAnalyticsWorkgroup",
       {
-        name: `GameAnalyticsWorkgroup-${cdk.Aws.STACK_NAME}`,
+        name: `${cdk.Aws.STACK_NAME}-Workgroup`,
         description: "Default workgroup for the solution workload",
         recursiveDeleteOption: true, // delete the associated queries when stack is deleted
         state: "ENABLED",
@@ -257,12 +84,6 @@ export class DataLakeConstruct extends Construct {
           },
         },
       }
-    );
-
-    this.createDefaultAthenaQueries(
-      gameEventsDatabase.ref,
-      props.config.RAW_EVENTS_TABLE,
-      `GameAnalyticsWorkgroup-${cdk.Aws.STACK_NAME}`
     );
 
     const cfnDataCatalogEncryptionSettings =
@@ -712,6 +533,7 @@ export class DataLakeConstruct extends Construct {
 
     this.gameEventsDatabase = gameEventsDatabase;
     this.rawEventsTable = rawEventsTable;
+    this.gameAnalyticsWorkgroup = gameAnalyticsWorkgroup;
 
     new cdk.CfnOutput(this, "GameEventsEtlJobOutput", {
       description:
